build: fix Makefile target "zip"
It was giving an error saying that couldn't compress two
files with the same name into the same archive (because
-j was being used, which prevents folder hierarchies to
be used inside the archive).

I guess I wanted to use -j before to not have the archived
be decompressed into a src/GWallet.yadda/yadda/ subfolder,
but now we cannot use it because there are resource files
in localization subfolders which are named the same. So
to avoid the stupid subfolder tree now we use SetCurrentDir()
to get inside, then make the zip, and then restore the
current dir as it was before.